Add inat tfrecode generator

+++ research/slim/datasets/dataset_utils.py
@@ -87,11 +87,12 @@ def download_and_uncompress_tarball(tarball_url, dataset_dir):
   filename = tarball_url.split('/')[-1]
   filepath = os.path.join(dataset_dir, filename)
 
-  def _progress(count, block_size, total_size):
-    sys.stdout.write('\r>> Downloading %s %.1f%%' % (
-        filename, float(count * block_size) / float(total_size) * 100.0))
+  if tf.gfile.Exists(filepath) == False:
+    def _progress(count, block_size, total_size):
+      sys.stdout.write('\r>> Downloading %s %.1f%%' % (
+          filename, float(count * block_size) / float(total_size) * 100.0))
     sys.stdout.flush()
-  filepath, _ = urllib.request.urlretrieve(tarball_url, filepath, _progress)
+    filepath, _ = urllib.request.urlretrieve(tarball_url, filepath, _progress)
   print()
   statinfo = os.stat(filepath)
   print('Successfully downloaded', filename, statinfo.st_size, 'bytes.')
